Moved "limitBodySize"
